add repair buildings

+++ src/main/java/models/MaterialInstance.java
@@ -2,9 +2,9 @@ package models;
 
 public class MaterialInstance {
     public final Material material;
-    public final double amount;
+    public final int amount;
 
-    public MaterialInstance(Material material, double amount) {
+    public MaterialInstance(Material material, int amount) {
         this.material = material;
         this.amount = amount;
     }
